adiciona lib de icones e nova menu bar
