+++ frontend/puppyfinder/src/pages/AdoptionListPage/fakeData.js
@@ -1,4 +1,4 @@
-const fakeDataAdoptionList = new Array(50).fill({
+const fakeDataAdoptionList = new Array(15).fill({
   nome: 'Nome do dog',
   localizacao: 'Centro - Belo Horizonte',
   descricao: 'Uma descrição sobre o dog e como ele foi encontrado - Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam tempor tortor at elit posuere, in vulputate turpis tincidunt. Cras non augue condimentum, blandit eros sed, facilisis leo. Morbi nunc velit, aliquet id erat eu, consectetur sodales enim. Morbi at nisi vitae ipsum dapibus viverra. Maecenas interdum eleifend odio, vitae faucibus nibh rutrum ut.',
