+++ frontend/puppyfinder/src/pages/AdoptionListPage/styles.js
@@ -1,0 +1,23 @@
+import {
+  makeStyles,
+} from '@material-ui/core';
+
+const useStyles = makeStyles(() => ({
+  card: {
+    height: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  cardMedia: {
+    paddingTop: '56.25%',
+  },
+  locationDiv: {
+    display: 'flex',
+    flexDirection: 'row',
+  },
+  locationIcon: {
+    fontSize: 18,
+  },
+}));
+
+export default useStyles;
