+++ frontend/puppyfinder/src/routes/index.jsx
@@ -6,9 +6,11 @@ import {
 } from 'react-router-dom';
 
 import AdoptionListPage from '../pages/AdoptionListPage';
+import MenuBar from '../layout/MenuBar';
 
 const Routes = () => (
   <Router>
+    <MenuBar />
     <Switch>
       <Route path="/" exact component={AdoptionListPage} />
     </Switch>
